rename goals to questions in routes, refactor questionsHelper
