+++ client/src/components/Application.js
@@ -3,11 +3,9 @@ import Greeting from "./Greeting";
 import Wall from "./Wall";
 import Rewards from "./Rewards/Index";
 import Insight from "./Rewards/Insight/Index";
-import HeroImage from "./HeroImage";
 import Navbar from "./LandingPage/Navbar";
 import QuestionList from "./QuestionList/Index";
 import Mood from "./Mood/Mood";
-import "./HeroImage.scss";
 import useApplicationData from "../hooks/useApplicationData";
 import { answeredGoals } from "../helpers/filterbyToday";
 import { Container } from "@material-ui/core";
