--- client/src/components/HeroImage.js
@@ -1,14 +1,0 @@
-import React from "react";
-import "./HeroImage.scss"
-
-export default function logoutPrompt(props){
-  
-  return (
-    <main  id="login-prompt">
-      
-      <h4 style={{color:'skyblue',fontWeight:'bold'}}>  </h4> 
-      <img src="images/logout-image2.png" id="logout-prompt-pic" />
-      
-    </main>
-  );
-}
