--- client/src/components/HeroImage.scss
@@ -1,18 +1,0 @@
-@import "../styles/variables.scss";
-
-#logout-prompt-pic{
-  height: 100%;
-  max-height: 60vh;
-  border-radius: 0px;
-  margin: 1em;
-}
-
-#logout-prompt{
-  padding: 1.2rem 1.5rem;
-  text-align: justify;
-  flex-flow: row wrap;
-  color: #fff;
-  background-color: $indezone-navy;
-  display: inline;
-  border: 10px double $white;
-}
